pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.1"
description = "Toolkit for ai usage."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx[socks]>=0.28.1",
    "ipython>=8.37.0",
    "openai>=2.14.0",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "typer-utils@git+https://github.com/qsoyq/typer_utils",
]

[project.scripts]
ai-assistant = "ai_assistant.commands.main:cmd"
ai-assistant-similar-questions = "ai_assistant.commands.similar_questions:cmd"

[tool.hatch.build.targets.wheel]
packages = ["ai_assistant"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
# Same as Black.
line-length = 200
indent-width = 4

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
# docstring-code-line-length = "dynamic"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
exclude = ['^venv/', '^__init__\.py$']
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "typer_utils.*",
]
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pytest.ini_options]
markers = []
